requirements.txt
python-amazon-paapi==5.0.1
APScheduler==3.10.4
requests_oauthlib==1.3.1
pyshorteners==1.0.1
python-dotenv==1.0.0
